server/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
google-api-python-client==2.118.0
google-generativeai==0.3.2
python-dotenv==1.0.1
pydantic==2.6.1
requests==2.31.0
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
asyncpg==0.29.0
